# api.yaml
swagger: '2.0'
info:
  title: My real estate api
  description: This is a real estate documentation made by me S13
  version: '0.1.0'
  contact:
    email: smdays13@gmail.com
  license:
    name: MIT
    url: https://spdx.org/licenses/MIT.html
  servers:
    -url: "http://localhost:4000/"
    description: this my local server
    -url: https://renbyexample.com
    description: this is my development server
schemes: [http,https]
...
paths:
...
  /users/{userId}:
    get:
      summary: Returns a user by ID.
      description: Returns the Specific User
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type : integer
            format: int64
      responses: 
        '200':
          description: OK
          schema:
            type: object
            example: {name: syed}

#Post json/body & Body required
  # /api/users:
  #   post:
  #     summary: Adds a new user
  #     produces:
  #       - application/json
  #     consumes:
  #       - application/json
  #     parameters:
  #       - in: body
  #         name: body
  #         schema:
  #           type: object
  #           required:
  #             - name
  #             - id
  #           properties:
  #             id:
  #               type: integer
  #             name:
  #               type: string
  #             email:
  #               type: string
  #             profileImg:
  #               type: string
  #             password:
  #               type: string
  #             passwordConfirm:
  #               type: string
  #     responses:
  #       '200':
  #         description: OK


#Defining Body Reference
  /api/users:
    post:
      summary: Adds a new user
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          schema:
           $ref: "#/components/schemas/User"
      responses:
        '200':
          description: OK


#Query
#Get Query
...
components:
  schemas:
    User:
      type: object
      required:
        - name
        - id
      properties:
        id:
          type: integer
        name:
          type: string